third_party/blink/web_tests/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-divisor-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'divisor' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('divisor') is "20"
PASS convolveMatrixElement.getAttribute('divisor') is "12"
PASS successfullyParsed is true

TEST COMPLETE
